+++ src/components/pages/index.js
@@ -6,3 +6,4 @@ export { default as CreationsPage } from './CreationsPage';
 export { default as Stocazzato } from './Stocazzato';
 export { default as WorkInProgress } from './WorkInProgress';
 export { default as LighthouseStats } from './LighthouseStats';
+export { default as TrophiesPage } from './TrophiesPage';